Introduce CPURegistersView and TraceWidget
