+++ src/gui/Src/Gui/RegistersView.h
@@ -127,7 +127,7 @@ public:
         }
     };
 
-    explicit RegistersView(CPUWidget* parent);
+    explicit RegistersView(QWidget* parent);
     ~RegistersView();
 
     QSize sizeHint() const;
@@ -136,12 +136,12 @@ public:
     static void operator delete(void* p);
     int getEstimateHeight();
 
+    void setRegisters(REGDUMP* reg); //This is really a protected member, but we need to put REGDUMP into it from TraceWidget
+
 public slots:
-    void refreshShortcutsSlot();
-    void updateRegistersSlot();
-    void displayCustomContextMenuSlot(QPoint pos);
-    void setRegister(REGISTER_NAME reg, duint value);
-    void debugStateChangedSlot(DBGSTATE state);
+    virtual void refreshShortcutsSlot();
+    virtual void displayCustomContextMenuSlot(QPoint pos);
+    virtual void debugStateChangedSlot(DBGSTATE state);
     void reload();
     void ShowFPU(bool set_showfpu);
     void onChangeFPUViewAction();
@@ -160,44 +160,15 @@ protected:
 
     // use-in-class-only methods
     void drawRegister(QPainter* p, REGISTER_NAME reg, char* value);
-    void setRegisters(REGDUMP* reg);
     char* registerValue(const REGDUMP* regd, const REGISTER_NAME reg);
     bool identifyRegister(const int y, const int x, REGISTER_NAME* clickedReg);
     QString helpRegister(REGISTER_NAME reg);
-    void CreateDumpNMenu(QMenu* dumpMenu);
 
-    void displayEditDialog();
     void ensureRegisterVisible(REGISTER_NAME reg);
 
 protected slots:
-    void fontsUpdatedSlot();
-    void onIncrementAction();
-    void onDecrementAction();
-    void onIncrementx87StackAction();
-    void onDecrementx87StackAction();
-    void onZeroAction();
-    void onSetToOneAction();
-    void onModifyAction();
-    void onToggleValueAction();
-    void onUndoAction();
-    void onCopyToClipboardAction();
-    void onCopyFloatingPointToClipboardAction();
-    void onCopySymbolToClipboardAction();
-    void onCopyAllAction();
-    void onFollowInDisassembly();
-    void onFollowInDump();
-    void onFollowInDumpN();
-    void onFollowInStack();
-    void onFollowInMemoryMap();
-    void onIncrementPtrSize();
-    void onDecrementPtrSize();
-    void onPushAction();
-    void onPopAction();
-    void onHighlightSlot();
     void InitMappings();
-    // switch SIMD display modes
-    void onSIMDMode();
-    void onFpuMode();
+    void fontsUpdatedSlot();
     QString getRegisterLabel(REGISTER_NAME);
     int CompareRegisters(const REGISTER_NAME reg_name, REGDUMP* regdump1, REGDUMP* regdump2);
     SIZE_T GetSizeRegister(const REGISTER_NAME reg_name);
@@ -210,14 +181,13 @@ protected slots:
     //unsigned int GetControlWordRCValueFromString(const char* string);
     QString GetMxCsrRCStateString(unsigned short);
     //unsigned int GetMxCsrRCValueFromString(const char* string);
-    void ModifyFields(const QString & title, STRING_VALUE_TABLE_t* table, SIZE_T size);
     //unsigned int GetStatusWordTOPValueFromString(const char* string);
     QString GetStatusWordTOPStateString(unsigned short state);
     void appendRegister(QString & text, REGISTER_NAME reg, const char* name64, const char* name32);
-    void disasmSelectionChangedSlot(dsint va);
-private:
+
+    void onCopyAllAction();
+protected:
     QPushButton* mChangeViewButton;
-    CPUWidget* mParent;
     bool mShowFpu;
     int mVScrollOffset;
     int mRowsNeeded;
@@ -267,47 +237,6 @@ private:
     unsigned int mRowHeight, mCharWidth;
     // SIMD registers display mode
     char mFpuMode; //0 = order by ST(X), 1 = order by x87rX, 2 = MMX registers
-    // context menu actions
-    QMenu* mSwitchSIMDDispMode;
-    QAction* mDisplaySTX;
-    QAction* mDisplayx87rX;
-    QAction* mDisplayMMX;
-    QAction* mFollowInDump;
-    QAction* wCM_Increment;
-    QAction* wCM_Decrement;
-    QAction* wCM_IncrementPtrSize;
-    QAction* wCM_DecrementPtrSize;
-    QAction* wCM_Push;
-    QAction* wCM_Pop;
-    QAction* wCM_Zero;
-    QAction* wCM_SetToOne;
-    QAction* wCM_Modify;
-    QAction* wCM_ToggleValue;
-    QAction* wCM_Undo;
-    QAction* wCM_CopyToClipboard;
-    QAction* wCM_CopyFloatingPointValueToClipboard;
-    QAction* wCM_CopySymbolToClipboard;
-    QAction* wCM_CopyAll;
-    QAction* wCM_FollowInDisassembly;
-    QAction* wCM_FollowInDump;
-    QAction* wCM_FollowInStack;
-    QAction* wCM_FollowInMemoryMap;
-    QAction* wCM_Incrementx87Stack;
-    QAction* wCM_Decrementx87Stack;
-    QAction* wCM_ChangeFPUView;
-    QAction* wCM_Highlight;
-    QAction* SIMDHex;
-    QAction* SIMDFloat;
-    QAction* SIMDDouble;
-    QAction* SIMDSWord;
-    QAction* SIMDUWord;
-    QAction* SIMDHWord;
-    QAction* SIMDSDWord;
-    QAction* SIMDUDWord;
-    QAction* SIMDHDWord;
-    QAction* SIMDSQWord;
-    QAction* SIMDUQWord;
-    QAction* SIMDHQWord;
     dsint mCip;
     std::vector<std::pair<const char*, uint8_t>> mHighlightRegs;
 };
