+++ src/gui/Src/Tracer/TraceBrowser.h
@@ -42,7 +42,6 @@ private:
     void setupRightClickContextMenu();
     void makeVisible(duint index);
     QString getAddrText(dsint cur_addr, char label[MAX_LABEL_SIZE], bool getLabel);
-    QString getIndexText(duint index) const;
     RichTextPainter::List getRichBytes(const Instruction_t & instr) const;
     void pushSelectionInto(bool copyBytes, QTextStream & stream, QTextStream* htmlStream = nullptr);
     void copySelectionSlot(bool copyBytes);
@@ -115,6 +114,7 @@ private:
 
 signals:
     void displayReferencesWidget();
+    void updateTraceRegistersView(void* registers);
 
 public slots:
     void openFileSlot();
