+++ src/x64dbg_translations.pro
@@ -225,7 +225,9 @@ SOURCES += \
     gui/Src/Gui/MultiItemsSelectWindow.cpp \
     gui/Src/BasicView/AbstractStdTable.cpp \
     gui/Src/Gui/ZehSymbolTable.cpp \
-    gui/Src/BasicView/StdSearchListView.cpp
+    gui/Src/BasicView/StdSearchListView.cpp \
+	gui/Src/Gui/CPURegistersView.cpp \
+	gui/Src/Tracer/TraceWidget.cpp
 
 HEADERS += \
     gui/Src/Exports.h \
@@ -462,7 +464,9 @@ HEADERS += \
     gui/Src/Gui/MultiItemsSelectWindow.h \
     gui/Src/BasicView/AbstractStdTable.h \
     gui/Src/Gui/ZehSymbolTable.h \
-    gui/Src/BasicView/StdSearchListView.h
+    gui/Src/BasicView/StdSearchListView.h \
+	gui/Src/Gui/CPURegistersView.h \
+	gui/Src/Tracer/TraceWidget.h
 
 FORMS += \
     gui/Src/Gui/AppearanceDialog.ui \
@@ -501,7 +505,8 @@ FORMS += \
     gui/Src/Gui/SimpleTraceDialog.ui \
     gui/Src/Gui/StructWidget.ui \
     gui/Src/Gui/MessagesBreakpoints.ui \
-    gui/Src/Gui/AboutDialog.ui
+    gui/Src/Gui/AboutDialog.ui \
+	gui/Src/Tracer/TraceWidget.ui
 
 TRANSLATIONS += \
     gui/Translations/x64dbg.ts
